requirements.txt
Flask
Flask-SocketIO
Eventlet
python-dotenv

